--- a/.env.default
+++ b/.env.default
@@ -29,8 +29,8 @@ LLM_GENERAL_MODEL_1=claude-3-7-sonnet-20250219
 LLM_GENERAL_MODEL_2=gpt-3.5-turbo
 
 # SQL-specific models (for SQL generation, database queries)
-LLM_SQL_MODEL_1=sqlcoder-7b-2
-LLM_SQL_MODEL_2=codellama-7b-instruct
+LLM_SQL_MODEL_1=codellama-7b-instruct
+LLM_SQL_MODEL_2=sqlcoder-7b-2
 LLM_SQL_MODEL_3=qwen/qwen3-coder-30b
 LLM_SQL_MODEL_4=gpt-oss-120b