--- a/.env.github
+++ b/.env.github
@@ -17,7 +17,6 @@ CLAUDE_ENDPOINT=https://api.anthropic.com
 
 # Local Model Configuration (LM Studio)
 LM_STUDIO_ENDPOINT=http://localhost:1234/v1
-LM_STUDIO_MODEL=openai/gpt-oss-20b
 
 # LLM Manager Configuration
 LLM_ROUTING_STRATEGY=preference