--- a/src/llm-manager/config.ts
+++ b/src/llm-manager/config.ts
@@ -92,14 +92,8 @@ const loadConfigFromEnv = (): Effect.Effect<LLMConfig, LLMError, never> =>
       }
     }
 
-    if (process.env.LM_STUDIO_MODEL) {
-      if (baseConfig.models.llama) {
-        baseConfig.models.llama = {
-          ...baseConfig.models.llama,
-          modelPath: process.env.LM_STUDIO_MODEL
-        }
-      }
-    }
+    // NOTE: LM_STUDIO_MODEL removed - use LLM_SQL_MODEL_* instead
+    // The default llama model will be configured through LLM_SQL_MODEL_1 if available
 
     // Load SQL-specific models from LLM_SQL_MODEL_* environment variables
     const sqlModelKeys = Object.keys(process.env).filter((key) => key.startsWith('LLM_SQL_MODEL_'))
@@ -258,7 +252,6 @@ export const ENV_DOCS = {
 
   // Local Model Configuration
   LM_STUDIO_ENDPOINT: 'LM Studio endpoint (default: http://localhost:1234/v1)',
-  LM_STUDIO_MODEL: 'Local model name (default: openai/gpt-oss-20b)',
 
   // Routing Configuration
   LLM_ROUTING_STRATEGY: 'Routing strategy: cost, performance, or balanced',