… .claude/settings.json
… debug-browser.html
… docker-compose.yaml
… notes/design/adr/adr-014-llm-management-library-evaluation.md
… notes/design/features/feature-002-phase1.md
… notes/design/features/feature-004-analysis.md
… notes/design/features/feature-004-clickhouse-ai-prompt-strategy.md
… src/llm-manager/README.md
… src/ui-generator/test/integration/sql-vs-gpt-query-comparison.test.ts
… src/ui-generator/test/unit/clickhouse-ai-prompt-comparison.test.ts
… src/ui-generator/test/unit/diagnostic-query-improvement.test.ts
… src/ui-generator/test/unit/trace-level-diagnostic.test.ts